Back Arrow goes back to existing parent (#979)
This change makes the preview back arrow take you back to a parent
directory that already exists if you are stuck in a bad state. this
prevents it from opening a directory that doesn't exist as a file, which
can happen if you use the typeahead to reach a file that doesn't exist
inside a directory that doesn't exist.