Upgrade Vite and other deps to resolve packaging issue (#1387)
There was a bug in Vite 6.0.1
(https://github.com/vitejs/vite/issues/18810) that was preventing NodeJS
libraries from being bundled in the main process. This meant that our
packaged app would refuse to run as it would be unable to find its Node
libraries.

This updates to Vite 6.0.2 to resolve this issue and also updates a few
other deps while we're at it.

Also removes the SASS modern-compiler config from our Vite config since
this is the default behavior in Vite 6.

closes #1373